Creating Subfolders for Graph Topics
