--- Graph/0-1 BFS based on SPOJ - KATHTHI.cpp
@@ -1,97 +1,0 @@
-tutorial link : https://codeforces.com/blog/entry/22276
-#pragma warning(disable:4786)
-#pragma warning(disable:4996)
-#include<bits/stdc++.h>
-#include <ext/pb_ds/assoc_container.hpp>
-#include <ext/pb_ds/trie_policy.hpp>
-#define pii pair<int,int>
-#define pll pair<long long ,long long>
-#define pli pair<long long , int>
-#define pil pair<int ,long long>
-#define pi acos(-1)
-#define pb push_back
-#define mkp make_pair
-#define all(a) a.begin(), a.end()
-#define IOS ios::sync_with_stdio(0); cin.tie(0); cout.tie(0);
-#define ll long long
-#define MAX 300005
-#define INF 0x3f3f3f3f
-template <class T> inline T bigmod(T p,T e,T M){ll ret = 1LL;for(; e > 0LL; e >>= 1LL){if(e & 1LL) ret = (ret * p) % M;p = (p * p) % M;}return (T)ret;}
-template <class T> inline T modinverse(T a,T M){return bigmod(a,M-2,M);}   // M is prime}
-
-using namespace std;
-using namespace __gnu_pbds;
-
-typedef trie<string,null_type,trie_string_access_traits<>,pat_trie_tag,trie_prefix_search_node_update>pref_trie;
-
-
-
-
-int main(){
-    IOS
-    //freopen("output.txt","w",stdout);
-    int t;
-    cin>>t;
-
-    for(int tt=0;tt<t;tt++){
-        int r,c;
-        cin>>r>>c;
-
-        char arr[r][c];
-
-        for(int i=0;i<r;i++){
-            for(int j=0;j<c;j++){
-                cin>>arr[i][j];
-            }
-        }
-        bool vis[r][c];
-        memset(vis,false,sizeof vis);
-
-        deque<pair<pair<int,int>,int> >d;
-        d.pb({{0,0},0});
-        int ans = 90000000;
-        while(!d.empty()){
-            auto u = d.front();
-            d.pop_front();
-            int x = u.first.first;
-            int y = u.first.second;
-            int lvl = u.second;
-            if(x==r-1 && y==c-1)ans = min(ans,lvl);
-            if(vis[x][y]==true)continue;
-            vis[x][y]=true;
-            if(x>0){
-                if(arr[x][y]!=arr[x-1][y]){
-                    d.pb({{x-1,y},lvl+1});
-                }
-                else{
-                    d.push_front({{x-1,y},lvl});
-                }
-            }
-            if(y>0){
-                if(arr[x][y]!=arr[x][y-1]){
-                    d.pb({{x,y-1},lvl+1});
-                }
-                else{
-                    d.push_front({{x,y-1},lvl});
-                }
-            }
-            if(x<r-1){
-                if(arr[x][y]!=arr[x+1][y]){
-                    d.pb({{x+1,y},lvl+1});
-                }
-                else{
-                    d.push_front({{x+1,y},lvl});
-                }
-            }
-            if(y<c-1){
-                if(arr[x][y]!=arr[x][y+1]){
-                    d.pb({{x,y+1},lvl+1});
-                }
-                else{
-                    d.push_front({{x,y+1},lvl});
-                }
-            }
-        }
-        cout<<ans<<"\n";
-    }
-}
