+++ Graph/Articulation vertex/Articulation Point.cpp
